Index Geo Coordinate topics directly (#633).
The internal indexing approach is changed: the Geo Coordinate topics are indexed directly, so that they become an active part of the R-tree build up by Neo4j Spatial. This supports index update and delete operations more natural.

NOTE: this module is still experimental. Only use it with a testing DB.

See #633.