Remove entries from geospatial index (#633).
Once a Geo Coordinate topic is deleted (e.g. when an Address is deleted) the corresponding entry is removed from the geospatial index.

NOTE: this module is still experimental. Only use it with a testing DB.

See #633.

-----------------------------------------------------------------------

CORRECT COMMIT MESSAGE FOR PREVIOUS COMMIT:

Update geospatial index (#633).

Once a Geo Coordinate topic changes (e.g. when an Address is changed) the geospatial index is updated accordingly.

See #633.